Refactor code into src package
