+++ Complete_Model_HPC_callrate_v5_45_24.py
@@ -25,6 +25,9 @@ from torch.nn.functional import softmax
 import torchmetrics
 import torch.nn as nn
 import matplotlib.pyplot as plt
+from src.data.dataset import SCDataset
+from src.models.gat import GAT
+from src.train import run_training
 
 print('package importing done except for sag_pool_custom')
 
@@ -238,88 +241,10 @@ def check_device(pytorch_object):
 
 #define the dataset
 #t1
-class SCDataset(Dataset):
-    def __init__(self, root, transform=None, pre_transform=None, pre_filter=None):
-        super(SCDataset, self).__init__(root, transform, pre_transform, pre_filter)
-        self.root = root  # add this if you want to specify the root directory while creating an object of your dataset
-        self.edge_index = global_edge_index #the edge index is defined here and passed to data objects as they are loaded
-        self.call_rate = self.calculate_call_rate_by_node().to(device) #this is the call rate for the entire dataset. it's a tensor of size [num_nodes]
-
-    @property
-    def raw_file_names(self):
-        # Since your files are named "noedgeidx_<value>.pt", we can generate this list using a list comprehension
-        return [f'noedgeidx_{i}.pt' for i in graph_idx_list]  # adjust the graph idx list above
-
-    @property
-    def processed_file_names(self):
-        # If you're not doing any preprocessing, you can just reuse your raw_file_names here
-        return self.raw_file_names
-
-    def len(self):
-        return len(self.raw_file_names)
-    
-    def get_without_call_rates(self, idx): #this is referenced only once by the call rate method during initialization.
-        data = torch.load(osp.join(self.root, self.raw_file_names[idx]))
-        # skip call_rates line here
-        data.edge_index = self.edge_index.to(device)
-        if data.x.device != device:
-            data = data.to(device)
-        return data
-    
-    def get(self, idx):
-        data = torch.load(osp.join(self.root, self.raw_file_names[idx]))
-        data = data.to(device)
-        call_rate = self.call_rate # retrieve the pre-computed call rates
-        data.x = torch.cat((data.x, call_rate.unsqueeze(-1)), dim=-1)  # Append call rate as a new feature
-        data.edge_index = self.edge_index.to(device)
-        if data.x.device != device:
-            data = data.to(device)
-        return data
-
-    
-    def calculate_call_rate_by_node(self):
-        graph_count = self.len()  # total number of graphs
-        node_nonzero_count = torch.zeros(self.get_without_call_rates(0).num_nodes, dtype=torch.float).to(device)  # non-zero count per node
-        if synth_data == False:
-            for idx in range(graph_count):
-                data = self.get_without_call_rates(idx)
-                if data.x.is_sparse:  
-                    # If sparse, directly add the number of non-zero values for each node
-                    node_indices = data.x._indices()[0]  # Get the indices of non-zero elements
-                    node_nonzero_count.index_add_(0, node_indices.to(device), torch.ones(node_indices.size(0), device=device)) #this line is the replacement for the one below it v2.15.13
-                    #node_nonzero_count.index_add_(0, node_indices, torch.ones(node_indices.size(0))).to(device) #device error in this line
-                else: 
-                    input(f"calculating call rate, but graph {idx} isn't sparse! ")
-            call_rate = node_nonzero_count / graph_count  # calculate call rate
-        else:
-            
-            ''' #this version assumes feature matrix has no one hot encoding (ie one value only)
-            # Calculate call rate for synthetic data, which is not sparse
-            for idx in range(graph_count):
-                data = self.get(idx)
-                # If dense, manually get the indices of non-zero elements
-                node_indices = (data.x != 0).nonzero(as_tuple=True)[0]
-                node_nonzero_count.index_add_(0, node_indices.to(device), torch.ones(node_indices.size(0), device=device))
-            '''
-            #this untested gpt version checks call rate on zero element of each node only (ie the true feature)
-            for idx in range(graph_count):
-                data = self.get_without_call_rates(idx)
-                # Check where this element is not equal to zero
-                node_indices = (data.x[:, 0] != 0).nonzero(as_tuple=True)[0] # the [:, 0] selects first column (ie the feature)
-                node_nonzero_count.index_add_(0, node_indices.to(device), torch.ones(node_indices.size(0), device=device))
-            
-            call_rate = node_nonzero_count / graph_count  # calculate call rate
-
-        #sparsify the call rate
-        non_zero_indices = call_rate.nonzero(as_tuple=False).t() # Get the indices in COO format
-        sparse_call_rate = torch.sparse_coo_tensor(indices=non_zero_indices, 
-                                               values=call_rate[call_rate.nonzero(as_tuple=True)], 
-                                               size=call_rate.size()).to(device)
-        return sparse_call_rate
 
 
 #model starts here
-dataset = SCDataset(root_dir)
+dataset = SCDataset(root_dir, global_edge_index, graph_idx_list, device, synth_data)
 
 print('dataset defined')
 print('calculating call rate')
@@ -364,45 +289,13 @@ for step, data in enumerate(train_loader):
 """
 
 #altnerative new model, lots of changes. let's see if it works better
-class GAT(torch.nn.Module):
-    def __init__(self, hidden_channels):
-        super(GAT, self).__init__()
-        torch.manual_seed(12345)
-        self.conv1 = GATConv(dataset.num_node_features, hidden_channels, heads=num_heads)
-        self.conv2 = GATConv(hidden_channels*num_heads, hidden_channels, heads=num_heads)
-        self.pool1 = SAGPoolingCustom(hidden_channels*num_heads, min_score=None, multiplier=1.0, ratio = pooling_keep_count)
-        self.lin = Linear(hidden_channels*num_heads, num_classes) #I had dataset.num_classes here, but it loads every data object at once to find out how many classes there are which caused GPU death. instead made a global variable num_classes 
-        self.batch_norm1 = torch.nn.BatchNorm1d(hidden_channels*num_heads)
-        self.batch_norm2 = torch.nn.BatchNorm1d(hidden_channels*num_heads)
-
-    def forward(self, x, edge_index, batch):
-        x = self.conv1(x, edge_index) 
-        x = x.relu()
-        #x = self.batch_norm1(x)
-        #x, edge_index, _, batch, perm, score, pre_drop_score_one = self.pool1(x, edge_index, None, batch) # pooling after the conv layer
-        x = self.conv2(x, edge_index)
-        x = x.relu()
-        #x = self.batch_norm2(x)
-        x, edge_index, _, batch, perm, score, pre_drop_score_two = self.pool1(x, edge_index, None, batch) # pooling after the conv layer
-
-        x = global_mean_pool(x, batch)
-        #x = F.dropout(x, p=0.5, training=self.training)
-        x = self.lin(x)
-
-        #pre_drop_score_one = softmax(pre_drop_score_one, dim=0) #softmax to normalize across graphs
-        #pre_drop_score_two = softmax(pre_drop_score_two, dim=0) #softmax to normalize across graphs
-        #pre_drop_score_two = abs(pre_drop_score_two)
-
-        #pre_drop_score = pre_drop_score_one * pre_drop_score_two #the product of the two scores is the total feature 
-        pre_drop_score = pre_drop_score_two #test redefinition for model tuning
-        return x, pre_drop_score
 
 print('GAT defined')
 
 
 
 #going from here to training for loop took an hour or two. no idea why
-model = GAT(hidden_channels=64)
+model = GAT(dataset.num_node_features, hidden_channels=64, num_heads=num_heads, pooling_keep_count=pooling_keep_count, num_classes=num_classes)
 
 #distribute across gpus
 if torch.cuda.device_count() > 1:
@@ -419,283 +312,5 @@ criterion = torch.nn.CrossEntropyLoss()
 
 print('model defined')
 
-def train():
-    model.train()
 
-    for data in train_loader:
-        #try sending everything to device
-        data.x = data.x.to(device)
-        data.edge_index = data.edge_index.to(device)
-        data.batch = data.batch.to(device)
-        data.y = data.y.to(device)         
-
-        #actual training stuff here 
-        out,pre_drop_score = model(data.x, data.edge_index, data.batch)  
-        loss = criterion(out, data.y) 
-        loss.backward()  
-        optimizer.step()  
-        optimizer.zero_grad()  
-
-predrop_by_epoch_list = []
-#alternative new test function
-def test(loader):
-    model.eval()
-    
-    metric_precision = torchmetrics.Precision(task = 'binary', average='macro').to(device)
-    metric_f1 = torchmetrics.F1Score(task = 'binary', average='macro', num_classes=2).to(device)
-
-    correct = 0
-    total = 0
-
-    for data in loader:
-        #try sending everything to device
-        data.x = data.x.to(device)
-        data.edge_index = data.edge_index.to(device)
-        data.batch = data.batch.to(device)
-        data.y = data.y.to(device)
-        
-        
-        out, pre_drop_score = model(data.x, data.edge_index, data.batch)
-        
-        #if it's the last epoch accumulate the final epoch pre_drop_scores.
-        if epoch == num_epochs: #ie if it's the last epoch. 
-            global pre_drop_score_sum  # Needed to modify global copy of pre_drop_score_sum
-            pre_drop_score_sum = pre_drop_score_sum.to(device) #this is probably redundant. 
-
-            #all scores are batched together in mega-feature matrix. must cut them up into <batchsize> graphs
-            batched_pre_drop_scores = pre_drop_score.detach()  # Detach the score from the current computation graph
-            num_graphs_in_batch = batched_pre_drop_scores.size()[0]/datafirst.num_nodes
-            
-            #check the num graphs in batch is a whole number, suggesting the correct number of nodes in each graph
-            assert (batched_pre_drop_scores.size()[0]/datafirst.num_nodes).is_integer(), f"num nodes in batch = {num_graphs_in_batch}, not evenly divisble by {datafirst.num_nodes}"
-            
-            #once checked, you need to turn it from float to int to satisfy view function. Can't do this before checking because it would conceal badly shaped tensors
-            num_graphs_in_batch = int(num_graphs_in_batch)
-
-            #re-shape the tensor to separate graphs by rows
-            unbatched_pre_drop_scores = batched_pre_drop_scores.view(num_graphs_in_batch, datafirst.num_nodes)
-            split_pre_drop_scores = torch.split(unbatched_pre_drop_scores, split_size_or_sections=1, dim=0)
-
-            #check again the shape is right, add to the score sum
-            for i, graph_pre_drop_score in enumerate(split_pre_drop_scores):
-                assert graph_pre_drop_score.shape == (1,datafirst.num_nodes), f"graph_pre_drop_score {i} has incorrect shape {graph_pre_drop_score.shape}"
-                pre_drop_score_sum += graph_pre_drop_score.squeeze() #the squeeze takes it from size [1,x] to size [x]. no data lost, but size match acheived. 
-
-            #check that all the values were used by checking the len of the split score list is equal to batch size
-            assert len(split_pre_drop_scores) == num_graphs_in_batch, f"Not all values from the original tensor were used, expected {batch_size} tensors, got {len(split_pre_drop_scores)}"
-
-        #if it's not the final epoch, calculate predrop sum for the epoch and add it to the list of dictionaries of predropscoresums by epoch
-        else:
-            global predrop_by_epoch_list
-            single_batch_scoresum = torch.zeros(dataset[0].num_nodes).to(device)
-            #all scores are batched together in mega-feature matrix. must cut them up into <batchsize> graphs
-            batched_pre_drop_scores = pre_drop_score.detach()  # Detach the score from the current computation graph
-            num_graphs_in_batch = batched_pre_drop_scores.size()[0]/datafirst.num_nodes
-            
-            #check the num graphs in batch is a whole number, suggesting the correct number of nodes in each graph
-            assert (batched_pre_drop_scores.size()[0]/datafirst.num_nodes).is_integer(), f"num nodes in batch = {num_graphs_in_batch}, not evenly divisble by {datafirst.num_nodes}"
-            
-            #this line replaces the commented code below
-            single_batch_scoresum = torch.sum(batched_pre_drop_scores.view(int(num_graphs_in_batch), datafirst.num_nodes), dim=0)
-           
-            """
-            #once checked, you need to turn it from float to int to satisfy view function. Can't do this before checking because it would conceal badly shaped tensors
-            num_graphs_in_batch = int(num_graphs_in_batch)
-
-            #re-shape the tensor to separate graphs by rows
-            unbatched_pre_drop_scores = batched_pre_drop_scores.view(num_graphs_in_batch, datafirst.num_nodes)
-            split_pre_drop_scores = torch.split(unbatched_pre_drop_scores, split_size_or_sections=1, dim=0)
-
-
-            #check again the shape is right, add to the score sum
-            for i, graph_pre_drop_score in enumerate(split_pre_drop_scores):
-                assert graph_pre_drop_score.shape == (1,datafirst.num_nodes), f"graph_pre_drop_score {i} has incorrect shape {graph_pre_drop_score.shape}"
-                single_batch_scoresum += graph_pre_drop_score.squeeze() #the squeeze takes it from size [1,x] to size [x]. no data lost, but size match acheived. 
-
-            #check that all the values were used by checking the len of the split score list is equal to batch size
-            assert len(split_pre_drop_scores) == num_graphs_in_batch, f"Not all values from the original tensor were used, expected {batch_size} tensors, got {len(split_pre_drop_scores)}"
-            """
-            predrop_by_epoch_list.append({epoch: single_batch_scoresum.cpu().numpy()}) #sent to cpu because numpy can't handle gpus. unknown if this will cause problems.
-
-
-
-
-        pred = out.argmax(dim=1)
-        correct += (pred == data.y).sum().item()
-        total += data.y.size(0)
-        
-        metric_precision.update(pred, data.y)
-        metric_f1.update(pred, data.y)
-
-    accuracy = correct / total
-    precision = metric_precision.compute()
-    f1 = metric_f1.compute()
-    return accuracy, precision, f1
-
-
-print('train and test defined')
-print('starting the epochs...')
-#eochs with 4 graphs takes a minute total with 4 cells, no self edge 50kb
-epoch_start_time = time.time()
-metrics_list = []
-for epoch in range(1, num_epochs+1):
-    train()
-    train_acc, train_prec, train_f1 = test(train_loader)
-    test_acc, test_prec, test_f1 = test(test_loader)
-    metrics = (f' \n Epoch: {epoch:03d} | epoch_time: {round(time.time()-epoch_start_time,1)} | Train Acc: {train_acc:.4f} | Train Prec: {train_prec:.4f} | Train F1: {train_f1:.4f} | Test Acc: {test_acc:.4f} | Test Prec: {test_prec:.4f} | Test F1: {test_f1:.4f}' )
-    print(metrics)
-    model_notes += metrics
-
-    metrics_list.append({
-        'Epoch': epoch,
-        'Train_Acc': train_acc,
-        'Train_Prec': train_prec.item(),
-        'Train_F1': train_f1.item(),
-        'Test_Acc': test_acc,
-        'Test_Prec': test_prec.item(),
-        'Test_F1': test_f1.item()
-    })
-
-    epoch_start_time = time.time()
-
-print('training complete, moving to saving results')
-
-# Function to generate the timestamp
-def get_timestamp():
-    now = datetime.datetime.now()
-    timestamp = now.strftime("%Y%m%d_%H%M%S")  # Format: YYYYMMDD_HHMMSS
-    return timestamp
-
-#set single timestamp for matching score and model names
-timestamp = get_timestamp()
-model_notes += f"timestamp : {timestamp}"
-
-
-#dense up the call rate for saving
-dense_node_call_rate = node_call_rate.to_dense()
-
-#make a pd dataframe of the scores
-# Create a DataFrame with node IDs and their corresponding scores
-model_results_df = pd.DataFrame({
-    'node_id': range(len(pre_drop_score_sum.cpu())),
-    'score': pre_drop_score_sum.cpu().numpy(),
-    'call_rate': dense_node_call_rate.cpu().numpy(),
-    #'weighted_score': pre_drop_score_sum.cpu().numpy() % node_call_rate.cpu().numpy() #look up how to do tensor division
-})
-
-#set the path where results will be saved. 
-if synth_data == True:
-    outputs_folder = "synth_models"
-else:
-    outputs_folder = "models"
-
-
-#if it's a true model, you add node metadata to results 
-if synth_data == False:
-    #for each node, we want to add the metadata
-    #this takes about 20 mins on the laptop: huge time sink. You could pre-make the metadata part of the df to speed it up
-    for node_id in model_results_df['node_id']:
-        metadata_series = id_to_node_metadata[node_id][1]
-        for column, value in metadata_series.items():
-            model_results_df.loc[model_results_df['node_id'] == node_id, column] = value
-    # Set the node_id as the index of the DataFrame
-    model_results_df.set_index('node_id', inplace=True)
-    #save results to a csv
-    model_results_df.to_csv(f"{outputs_folder}/Model_{timestamp}_results.csv", index = False)
-else:  #save model results without appending node metadata
-    #save results to a csv
-    model_results_df.to_csv(f"{outputs_folder}/Model_{timestamp}_results.csv")
-
-#save call rate to a pt
-call_rate_filename = f"{outputs_folder}/Model_{timestamp}_call_rate.pt"
-torch.save(dense_node_call_rate.cpu(), call_rate_filename)
-
-#save notes to a txt file
-with open(f"{outputs_folder}/Model_{timestamp}_notes.txt",'w') as notes_file:
-    notes_file.write(model_notes)
-
-# Create an metrics by epoch DataFrame from list
-metrics_by_epoch_df = pd.DataFrame(metrics_list)
-# Save metrics by epoch to csv
-metrics_by_epoch_filename = f"{outputs_folder}/Model_{timestamp}_metrics_by_epoch.csv"
-metrics_by_epoch_df.to_csv(metrics_by_epoch_filename, index=False)
-
-#plot the training metrics by epoch
-metrics_by_epoch_df.plot(x='Epoch', y=['Train_Acc', 'Train_Prec', 'Train_F1'],kind='line', title='Training Performance', grid=True)
-plt.xlabel('Epoch')
-plt.ylabel('Performance')
-plt.savefig(f"{outputs_folder}/Model_{timestamp}_train_metrics_by_epoch.png",dpi=300)
-plt.clf()
-
-#plot the test metrics by epoch
-metrics_by_epoch_df.plot(x='Epoch', y=['Test_Acc', 'Test_Prec', 'Test_F1'],kind='line', title='Test Performance', grid=True)
-plt.xlabel('Epoch')
-plt.ylabel('Performance')
-plt.savefig(f"{outputs_folder}/Model_{timestamp}_test_metrics_by_epoch.png",dpi=300)
-plt.clf()
-
-#train vs test by epoch
-metrics_by_epoch_df.plot(x='Epoch', y=['Test_F1', 'Train_F1'],kind='line', title='Test vs Train F1', grid=True)
-plt.xlabel('Epoch')
-plt.ylabel('Performance')
-plt.savefig(f"{outputs_folder}/Model_{timestamp}_train_test_loss_by_epoch.png",dpi=300)
-plt.clf()
-
-#aggreaget the predrop by epoch scores to be summed across epochs, not batches
-# Initialize an aggregation dictionary
-aggregated_tensors = {}
-
-# Loop and aggregate
-for d in predrop_by_epoch_list:
-    for key, tensor in d.items():
-        if key in aggregated_tensors:
-            # Add to the existing tensor
-            aggregated_tensors[key] += tensor
-        else:
-            # Create a new entry
-            aggregated_tensors[key] = tensor
-
-# Convert the aggregation dictionary to the desired list of dictionaries format
-predrop_by_epoch_list = [{k: v} for k, v in aggregated_tensors.items()]
-
-
-
-#create a predropbyepoch dataframe from list
-#start by swapping the list of dictionaries to a dictionary of lists
-# Convert list of dictionaries into dictionary of lists
-dict_of_lists = {}
-for d in predrop_by_epoch_list:
-    for key, value in d.items():
-        if key not in dict_of_lists:
-            dict_of_lists[key] = []
-        dict_of_lists[key].extend(value)
-
-predrop_by_epoch = dict_of_lists
-predrop_by_epoch_df = pd.DataFrame(predrop_by_epoch)
-#save to csv
-predrop_by_epoch_filename = f"{outputs_folder}/Model_{timestamp}_predrop_scores_by_epoch.csv"
-predrop_by_epoch_df.to_csv(predrop_by_epoch_filename, index=False)
-
-#plot the predrop by epoch
-predrop_by_epoch_df.T.plot(figsize=(10, 6))
-plt.title('Scores by Epoch')
-plt.xlabel('Epoch')
-plt.ylabel('Epoch Score Sum')
-plt.legend(title='Nodes', loc='upper right', labels=predrop_by_epoch_df.columns)
-plt.grid(True)
-plt.tight_layout()
-plt.savefig(f"{outputs_folder}/Model_{timestamp}_scores_by_epoch.png",dpi=300)
-plt.clf()
-
-
-#save model
-model_filename = f"{outputs_folder}/Model_{timestamp}.pth"
-torch.save(model.state_dict(), model_filename)
-fin_string = f"finished. total run time: {time.time() - script_start_time} | everything saved but notes. Now saving notes..."
-print(fin_string)
-model_notes += f"finished. total run time: {time.time() - script_start_time} | everything saved but notes. Now saving notes..."
-
-print('model run finished, printing model notes')
-print(f'{model_notes}')
-
-pass
+run_training()
